+++ src/app/user-module/navbar/navbar.component.ts
@@ -17,8 +17,10 @@ export class NavbarComponent implements OnInit {
   }
 
   logout(){
+    if (confirm("Are you sure you want to logout?")) {
     localStorage.clear();
     this.router.navigate(['/login']);
+    }
   }
 
 }
